fixed check_permissions(), created nested view for book in rent

+++ resources/rents.py
@@ -1,8 +1,11 @@
+from flask import jsonify
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+import flask_jwt_extended as jwt
+from sqlalchemy.orm import joinedload
+
 from db import db
-from models import BookRentModel, BookModel, ClientModel
+from models import BookRentModel, BookModel, check_admins_permissions
 from schema import PlainRentSchema, UpdateRentSchema, UpdateBookSchema
 from sqlalchemy.exc import SQLAlchemyError
 
@@ -12,23 +15,26 @@ blp = Blueprint("BookRental", __name__, description="Book rental CRUD.")
 @blp.route('/rental/<string:rent_id>')
 class BookRental(MethodView):
 
-    # @jwt_required()
+    @jwt.jwt_required()
     @blp.response(200, PlainRentSchema)
     def get(self, rent_id):
-        book_rent = BookRentModel.query.get_or_404(rent_id)
-        return book_rent
+        if check_admins_permissions(jwt.get_jwt_identity()):
+            book_rent = BookRentModel.query.filter_by(id=rent_id). \
+                options(joinedload(BookRentModel.books),
+                        joinedload(BookRentModel.client)).first()
+            return book_rent
 
-    # @jwt_required()
+        return jsonify({"message": "Lack of permissions."})
+
+    @jwt.jwt_required()
     def delete(self, rent_id):
-        book_rent = BookRentModel.query.get_or_404(rent_id)
-        # book = BookModel.
-        try:
+        if check_admins_permissions(jwt.get_jwt_identity()):
+            book_rent = BookRentModel.query.get_or_404(rent_id)
             db.session.delete(book_rent)
             db.session.commit()
-        except SQLAlchemyError:
-            abort(500, "SQLAlchemyError occurred while deleting BookRental from db.")
+            return jsonify({"message": f"Book {book_rent.id} deleted."})
+        return jsonify({"message": "Lack of permissions."})
 
-        return {"message": "Rent {} successfully deleted".format(rent_id)}
 
     @blp.arguments(UpdateRentSchema)
     @blp.response(201, PlainRentSchema)
@@ -49,14 +55,16 @@ class BookRental(MethodView):
 
         return book_rent
 
+
 @blp.route('/rental')
 class ListBookRental(MethodView):
 
     @blp.response(200, PlainRentSchema(many=True))
     def get(self):
-        return BookRentModel.query.all()
+        return BookRentModel.query.options(joinedload(BookRentModel.books), joinedload(BookRentModel.client))
 
-    # @jwt_required()
+    #todo: enable Regular create rents (ony for themselves) with book.title and book.author
+    # @jwt.jwt_required()
     @blp.arguments(PlainRentSchema)
     @blp.response(201, PlainRentSchema)
     def post(self, rent_data):
@@ -81,7 +89,8 @@ class ListBookRental(MethodView):
 @blp.route("/rental/client/<int:client_id>")
 class ListClientRents(MethodView):
 
-    #todo: modify so that get accepts email in json instead of client id
+    # todo: modify so that get accepts email in json instead of client id
+    # todo: Client can only see all its book_rents
     # @jwt_required()
     @blp.response(200, PlainRentSchema(many=True))
     def get(self, client_id):
